Refs #29802. Rename Spectra header file.
