+++ qt/widgets/common/inc/MantidQtWidgets/Common/FunctionModelDataset.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include "DllOption.h"
-#include "MantidQtWidgets/Common/Spectra.h"
+#include "MantidQtWidgets/Common/FunctionModelSpectra.h"
 
 #include <QList>
 #include <QString>
