+++ qt/widgets/common/inc/MantidQtWidgets/Common/FunctionModelSpectra.h
