+++ qt/widgets/common/src/FunctionModelSpectra.cpp
@@ -4,7 +4,7 @@
 //   NScD Oak Ridge National Laboratory, European Spallation Source,
 //   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 // SPDX - License - Identifier: GPL - 3.0 +
-#include "MantidQtWidgets/Common/Spectra.h"
+#include "MantidQtWidgets/Common/FunctionModelSpectra.h"
 #include "MantidKernel/Strings.h"
 
 #include <numeric>
